Refactor contact and account operations
