+++ packages/web-service/src/pages/contact/common/user/user.js
@@ -1,5 +1,6 @@
 import { APIRequests } from '../../../../services/api-requests.js'
-import { accountsRoute, contactOperations, contactsRoute } from '../common.js'
+import { accountsRoute, contactsRoute } from '../common.js'
+import { contactOperations } from '../operations.js'
 
 export const getUserData = _contactRole => async request => {
   const journeyData = await request.cache().getData()
